[BUGFIX] [PERSISTENCE] [@jnielavitzky] Fixed alot of stuff.

+++ persistence/src/main/java/edu/itba/paw/jimi/persistence/DishHibernateDao.java
@@ -2,17 +2,12 @@ package edu.itba.paw.jimi.persistence;
 
 import edu.itba.paw.jimi.interfaces.daos.DishDao;
 import edu.itba.paw.jimi.models.Dish;
-import org.springframework.beans.factory.annotation.Qualifier;
 import org.springframework.stereotype.Repository;
-import org.springframework.transaction.annotation.EnableTransactionManagement;
-import org.springframework.transaction.annotation.Transactional;
 
 import javax.persistence.EntityManager;
 import javax.persistence.PersistenceContext;
-import javax.persistence.PersistenceContextType;
 import javax.persistence.TypedQuery;
 import java.util.Collection;
-import java.util.List;
 
 
 @Repository
@@ -35,7 +30,7 @@ public class DishHibernateDao implements DishDao {
     }
 
     public Dish findById(long id) {
-        return em.find(Dish.class, id);
+        return em.find(Dish.class, (int)id);
     }
 
     public Collection<Dish> findAll() {
